refactor(VtkThreeView): Switch to vue 3 impl
